@@ -13,13 +13,12 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=5.4.0",
-        "symfony/symfony": "2.7.*",
-        "doctrine/orm": "~2.2,>=2.2.3,<2.5",
-        "doctrine/dbal": "<2.5",
-        "doctrine/doctrine-bundle": "~1.4",
+        "php": ">=5.5.9",
+        "symfony/symfony": "3.0.*",
+        "twig/twig": "~1.23|~2.0",
+        "doctrine/doctrine-bundle": "~1.5",
+        "doctrine/orm": "~2.4",
         "symfony/assetic-bundle": "~2.3",
-        "twig/twig": "~1.12",
         "xideapl/user": "dev-master@dev",
         "xideapl/base-bundle": "dev-master@dev"
     },
